make Scenario a final class
